% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound_functions.R
\name{load_audio}
\alias{load_audio}
\title{Load an audio file from disk}
\usage{
load_audio(file)
}
\arguments{
\item{file}{A character string specifying the path to the `.wav` file.}
}
\value{
A Wave object containing the audio signal.
}
\description{
This function reads a `.wav` audio file and converts it into a Wave object
that can be further processed. It is typically the first step in the
sound classification pipeline.
}
\details{
The function uses \code{tuneR::readWave()} to load the file.
It also checks whether the file exists before attempting to read it.
}
\examples{
\dontrun{
audio <- load_audio("bird1.wav")
}

}
