% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound_functions.R
\name{plot_spectrogram}
\alias{plot_spectrogram}
\title{Plot spectrogram of an audio file}
\usage{
plot_spectrogram(file)
}
\arguments{
\item{file}{A character string specifying the path to the audio file.}
}
\value{
A spectrogram plot.
}
\description{
This function visualizes the frequency content of an audio signal
over time using a spectrogram.
}
\details{
The spectrogram is generated using \code{seewave::spectro()}.
It shows how frequency components vary over time.
}
\examples{
\dontrun{
plot_spectrogram("bird1.wav")
}

}
